move event subscription to apps
